Add files via upload
aims to solve the challenges related to land transactions by digitizing records and keeping the ledger immutable and secure. Users can sign into  DigiEstate and register their lands, which are verified by the Lekhpal. The Tehsildar is the only role that can approve the transfer of ownership of lands from one user to another. The contract deployer is the SDM, who can be transferred if needed.